Re-add vehicle machines (rockets, RCS, ...), links, and signals
