+++ src/testapp/sessions/magnum.cpp
@@ -563,165 +563,4 @@ Session setup_shader_phong(
 } // setup_shader_phong
 
 
-
-#if 0
-
-struct IndicatorMesh
-{
-    Magnum::Color4 m_color;
-    MeshIdOwner_t m_mesh;
-};
-
-
-Session setup_thrust_indicators(
-        TopTaskBuilder&             rBuilder,
-        ArrayView<entt::any> const  topData,
-        Session const&              magnum,
-        Session const&              commonScene,
-        Session const&              parts,
-        Session const&              signalsFloat,
-        Session const&              scnRender,
-        Session const&              cameraCtrl,
-        Session const&              shFlat,
-        TopDataId const             idResources,
-        PkgId const                 pkg)
-{
-
-
-    using namespace osp::link;
-    using adera::gc_mtMagicRocket;
-    using adera::ports_magicrocket::gc_throttleIn;
-    using adera::ports_magicrocket::gc_multiplierIn;
-
-    static constexpr float indicatorScale = 0.0001f;
-
-    OSP_SESSION_UNPACK_TAGS(magnum,         TESTAPP_APP_MAGNUM);
-    OSP_SESSION_UNPACK_DATA(magnum,         TESTAPP_APP_MAGNUM);
-    OSP_SESSION_UNPACK_TAGS(commonScene,      TESTAPP_COMMON_SCENE);
-    OSP_SESSION_UNPACK_DATA(commonScene,      TESTAPP_COMMON_SCENE);
-    OSP_SESSION_UNPACK_TAGS(parts,          TESTAPP_PARTS);
-    OSP_SESSION_UNPACK_DATA(parts,          TESTAPP_PARTS);
-    OSP_SESSION_UNPACK_DATA(signalsFloat,   TESTAPP_SIGNALS_FLOAT)
-    OSP_SESSION_UNPACK_TAGS(signalsFloat,   TESTAPP_SIGNALS_FLOAT);
-    OSP_SESSION_UNPACK_TAGS(scnRender,      TESTAPP_COMMON_RENDERER);
-    OSP_SESSION_UNPACK_DATA(scnRender,      TESTAPP_COMMON_RENDERER);
-
-    OSP_SESSION_UNPACK_DATA(cameraCtrl,     TESTAPP_CAMERA_CTRL);
-    OSP_SESSION_UNPACK_TAGS(cameraCtrl,     TESTAPP_CAMERA_CTRL);
-    OSP_SESSION_UNPACK_DATA(shFlat,         TESTAPP_SHADER_FLAT);
-
-    auto &rResources    = top_get< Resources >      (topData, idResources);
-    auto &rBasic        = top_get< ACtxBasic >      (topData, idBasic);
-    auto &rActiveIds    = top_get< ActiveReg_t >    (topData, idActiveIds);
-    auto &rDrawing      = top_get< ACtxDrawing >    (topData, idDrawing);
-    auto &rDrawingRes   = top_get< ACtxDrawingRes > (topData, idDrawingRes);
-
-    Session thrustIndicator;
-#if 0 // RENDERENT
-    OSP_SESSION_ACQUIRE_DATA(thrustIndicator, topData, TESTAPP_INDICATOR);
-    thrustIndicator.m_tgCleanupEvt = tgCleanupMagnumEvt;
-
-    auto &rIndicator   = top_emplace<IndicatorMesh>(topData, idIndicator);
-    rIndicator.m_color = { 1.0f, 0.0f, 0.0f, 1.0f };
-    rIndicator.m_mesh  = SysRender::add_drawable_mesh(rDrawing, rDrawingRes, rResources, pkg, "cone");
-
-    thrustIndicator.task() = rBuilder.task().assign({tgRenderEvt, tgGlUse, tgDrawTransformReq, tgBindFboReq, tgFwdRenderMod, tgCamCtrlReq}).data(
-            "Render cursor",
-            TopDataIds_t{          idRenderGl,              idCamera,                      idDrawingRes,                         idScnRender,                 idScnParts,                             idSigValFloat,              idDrawShFlat,                            idCamCtrl,               idIndicator},
-            wrap_args([] (RenderGL& rRenderGl, Camera const& rCamera, ACtxDrawingRes const& rDrawingRes, ACtxSceneRenderGL const& rScnRender, ACtxParts const& rScnParts, SignalValues_t<float> const& rSigValFloat, ACtxDrawFlat& rDrawShFlat, ACtxCameraController const& rCamCtrl, IndicatorMesh& rIndicator) noexcept
-    {
-
-        ResId const     indicatorResId      = rDrawingRes.m_meshToRes.at(rIndicator.m_mesh.value());
-        MeshGlId const  indicatorMeshGlId   = rRenderGl.m_resToMesh.at(indicatorResId);
-        Mesh&           rIndicatorMeshGl    = rRenderGl.m_meshGl.get(indicatorMeshGlId);
-
-        ViewProjMatrix viewProj{rCamera.m_transform.inverted(), rCamera.perspective()};
-
-        //auto const matrix = viewProj.m_viewProj * Matrix4::translation(rCamCtrl.m_target.value());
-
-        PerMachType const& rockets = rScnParts.m_machines.m_perType[gc_mtMagicRocket];
-        Nodes const& floats = rScnParts.m_nodePerType[gc_ntSigFloat];
-
-        for (MachLocalId const localId : rockets.m_localIds.bitview().zeros())
-        {
-            MachAnyId const anyId           = rockets.m_localToAny[localId];
-            PartId const    part            = rScnParts.m_machineToPart[anyId];
-            ActiveEnt const partEnt         = rScnParts.m_partToActive[part];
-
-            auto const&     portSpan        = floats.m_machToNode[anyId];
-            NodeId const    throttleIn      = connected_node(portSpan, gc_throttleIn.m_port);
-            NodeId const    multiplierIn    = connected_node(portSpan, gc_multiplierIn.m_port);
-
-            float const     throttle        = std::clamp(rSigValFloat[throttleIn], 0.0f, 1.0f);
-            float const     multiplier      = rSigValFloat[multiplierIn];
-            float const     thrustMag       = throttle * multiplier;
-
-            if (thrustMag == 0.0f)
-            {
-                continue;
-            }
-
-            Matrix4 const& rocketDrawTf = rScnRender.m_drawTransform.get(partEnt);
-
-            auto const& matrix = Matrix4::from(rocketDrawTf.rotationNormalized(), rocketDrawTf.translation())
-                               * Matrix4::scaling({1.0f, 1.0f, thrustMag * indicatorScale})
-                               * Matrix4::translation({0.0f, 0.0f, -1.0f})
-                               * Matrix4::scaling({0.2f, 0.2f, 1.0f});
-
-            rDrawShFlat.m_shaderUntextured
-                .setColor(rIndicator.m_color)
-                .setTransformationProjectionMatrix(viewProj.m_viewProj * matrix)
-                .draw(rIndicatorMeshGl);
-        }
-    }));
-
-    thrustIndicator.task() = rBuilder.task().assign({tgCleanupMagnumEvt}).data(
-            "Clean up thrust indicator resource owners",
-            TopDataIds_t{             idDrawing,               idIndicator},
-            wrap_args([] (ACtxDrawing& rDrawing, IndicatorMesh& rIndicator) noexcept
-    {
-        rDrawing.m_meshRefCounts.ref_release(std::move(rIndicator.m_mesh));
-    }));
-
-    // Draw transforms in part entities are required for drawing indicators.
-    // This solution of adding them here is a bit janky but it works.
-
-    thrustIndicator.task() = rBuilder.task().assign({tgSyncEvt, tgPartReq}).data(
-            "Add draw transforms to rocket entities",
-            TopDataIds_t{                 idScnParts,                   idScnRender},
-            wrap_args([] (ACtxParts const& rScnParts, ACtxSceneRenderGL& rScnRender) noexcept
-    {
-        for (PartId const part : rScnParts.m_partDirty)
-        {
-            // TODO: maybe first check if the part contains any rocket machines
-
-            ActiveEnt const ent = rScnParts.m_partToActive[part];
-            if ( ! rScnRender.m_drawTransform.contains(ent) )
-            {
-                rScnRender.m_drawTransform.emplace(ent);
-            }
-        }
-    }));
-
-    // Called when scene is reopened
-    thrustIndicator.task() = rBuilder.task().assign({tgResyncEvt, tgPartReq}).data(
-            "Add draw transforms to all rocket entities",
-            TopDataIds_t{                 idScnParts,                   idScnRender},
-            wrap_args([] (ACtxParts const& rScnParts, ACtxSceneRenderGL& rScnRender) noexcept
-    {
-        for (PartId const part : rScnParts.m_partIds.bitview().zeros())
-        {
-            ActiveEnt const ent = rScnParts.m_partToActive[part];
-            if ( ! rScnRender.m_drawTransform.contains(ent) )
-            {
-                rScnRender.m_drawTransform.emplace(ent);
-            }
-        }
-    }));
-#endif
-    return thrustIndicator;
-}
-
-#endif
-
 } // namespace testapp::scenes
